+++ astro.config.mjs
@@ -1,12 +1,9 @@
 // @ts-check
 import { defineConfig } from 'astro/config';
-import tailwind from '@astrojs/tailwind';
-import cloudflare from "@astrojs/cloudflare";
+import tailwind from "@astrojs/tailwind";
 
 // https://astro.build/config
 export default defineConfig({
-  output: "server",
-  adapter: cloudflare(),
   integrations: [tailwind()],
   vite: {
     define: {